Converted int to size_t.

+++ blockpad.c
@@ -41,7 +41,7 @@
 # include <string.h>
 #endif
 
-memchunk* pkcs5Pad(int blockbytes, memchunk* tmp)
+memchunk* pkcs5Pad(size_t blockbytes, memchunk* tmp)
 {
 	if (tmp)
 	{
@@ -56,7 +56,7 @@ memchunk* pkcs5Pad(int blockbytes, memchunk* tmp)
 	return tmp;
 }
 
-memchunk* pkcs5Unpad(int blockbytes, memchunk* tmp)
+memchunk* pkcs5Unpad(size_t blockbytes, memchunk* tmp)
 {
 	if (tmp)
 	{
@@ -80,7 +80,7 @@ memchunk* pkcs5Unpad(int blockbytes, memchunk* tmp)
 	return tmp;
 }
 
-memchunk* pkcs5PadCopy(int blockbytes, const memchunk* src)
+memchunk* pkcs5PadCopy(size_t blockbytes, const memchunk* src)
 {
 	memchunk* tmp;
 	byte padvalue = blockbytes - (src->size % blockbytes);
@@ -99,7 +99,7 @@ memchunk* pkcs5PadCopy(int blockbytes, const memchunk* src)
 	return tmp;
 }
 
-memchunk* pkcs5UnpadCopy(int blockbytes, const memchunk* src)
+memchunk* pkcs5UnpadCopy(size_t blockbytes, const memchunk* src)
 {
 	memchunk* tmp;
 	byte padvalue;
